+++ Modules/Combat/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.1.8
 
+### Changed
+
+* Default auto-selection border color changed from Magenta to Aqua.
+
 ### Fixed
 
 * Fixed possible out-of-bounds exception in auto-selection draw.